[feat] Move to models folder
